requirements.txt
django
psycopg2
djangorestframework
django-rest-swagger
django-oauth2-provider
django-filter
django-simple-history
